Refactor Shogi module structure
